# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Portal API",
    "version": "1.0.0",
    "description": "Buy, sell and rent property across Myanmar. Prices are decimal strings in the smallest currency unit (MMK has no subunit), because a kyat sale price can exceed the range where a JSON number stays exact."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/listings/{id}/similar": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Listing id or public reference"
        }
      ],
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Comparable listings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "publicRef": {
                            "type": "string"
                          },
                          "dealType": {
                            "type": "string",
                            "enum": [
                              "SALE",
                              "RENT"
                            ]
                          },
                          "status": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "price": {
                            "type": "object",
                            "properties": {
                              "amount": {
                                "type": [
                                  "string",
                                  "null"
                                ],
                                "pattern": "^\\d+$"
                              },
                              "currency": {
                                "type": "string",
                                "enum": [
                                  "MMK",
                                  "USD"
                                ]
                              },
                              "lakhLabel": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "isNegotiable": {
                                "type": "boolean"
                              },
                              "onRequest": {
                                "type": "boolean"
                              },
                              "rentPeriod": {
                                "type": [
                                  "string",
                                  "null"
                                ],
                                "enum": [
                                  "MONTHLY",
                                  "YEARLY",
                                  null
                                ]
                              }
                            }
                          },
                          "category": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              },
                              "nameEn": {
                                "type": "string"
                              },
                              "nameMy": {
                                "type": "string"
                              }
                            }
                          },
                          "township": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              },
                              "nameEn": {
                                "type": "string"
                              },
                              "nameMy": {
                                "type": "string"
                              }
                            }
                          },
                          "coverImage": {
                            "type": [
                              "object",
                              "null"
                            ],
                            "properties": {
                              "url": {
                                "type": "string"
                              },
                              "thumbUrl": {
                                "type": "string"
                              }
                            }
                          },
                          "imageCount": {
                            "type": "integer"
                          },
                          "isFeatured": {
                            "type": "boolean"
                          },
                          "publishedAt": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "viewCount": {
                            "type": "integer"
                          },
                          "bedrooms": {
                            "type": "integer"
                          },
                          "bathrooms": {
                            "type": "integer"
                          },
                          "floorAreaSqft": {
                            "type": "integer"
                          },
                          "landAreaSqft": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/listings/{id}/view": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Listing id or public reference"
        }
      ],
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Record a view",
        "responses": {
          "204": {
            "description": "Counted"
          }
        }
      }
    },
    "/listings/{id}/submit": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Listing id"
        }
      ],
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Submit for review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "publicRef": {
                      "type": "string"
                    },
                    "dealType": {
                      "type": "string",
                      "enum": [
                        "SALE",
                        "RENT"
                      ]
                    },
                    "status": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "price": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "pattern": "^\\d+$"
                        },
                        "currency": {
                          "type": "string",
                          "enum": [
                            "MMK",
                            "USD"
                          ]
                        },
                        "lakhLabel": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "isNegotiable": {
                          "type": "boolean"
                        },
                        "onRequest": {
                          "type": "boolean"
                        },
                        "rentPeriod": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "enum": [
                            "MONTHLY",
                            "YEARLY",
                            null
                          ]
                        }
                      }
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "nameEn": {
                          "type": "string"
                        },
                        "nameMy": {
                          "type": "string"
                        }
                      }
                    },
                    "township": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "nameEn": {
                          "type": "string"
                        },
                        "nameMy": {
                          "type": "string"
                        }
                      }
                    },
                    "coverImage": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "url": {
                          "type": "string"
                        },
                        "thumbUrl": {
                          "type": "string"
                        }
                      }
                    },
                    "imageCount": {
                      "type": "integer"
                    },
                    "isFeatured": {
                      "type": "boolean"
                    },
                    "publishedAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "viewCount": {
                      "type": "integer"
                    },
                    "bedrooms": {
                      "type": "integer"
                    },
                    "bathrooms": {
                      "type": "integer"
                    },
                    "floorAreaSqft": {
                      "type": "integer"
                    },
                    "landAreaSqft": {
                      "type": "integer"
                    },
                    "description": {
                      "type": "string"
                    },
                    "address": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "addressHidden": {
                      "type": "boolean"
                    },
                    "attributes": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "amenities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "nameEn": {
                            "type": "string"
                          },
                          "nameMy": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "thumbUrl": {
                            "type": "string"
                          },
                          "width": {
                            "type": "integer"
                          },
                          "height": {
                            "type": "integer"
                          },
                          "isCover": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "contact": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "viber": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    },
                    "enquiryCount": {
                      "type": "integer"
                    },
                    "expiresAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "rejectionReason": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/listings/{id}/renew": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Listing id"
        }
      ],
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Extend the expiry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Renewed"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/me/listings": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "The caller’s own listings, with status counts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}